requirements.txt
ultralytics>=8.0.0
opencv-python
onnx
onnxruntime-gpu
numpy
torch
torchvision
tensorrt
fastapi
uvicorn[standard]
python-multipart
pydantic
pyyaml
python-dotenv
redis
prometheus-client
streamlit
plotly
websockets
pillow
requests
